Fix slow `snapper list` calls
`snapper list` is slow when qgroups are enabled. This is because by
default it lists used space ("management wanted it on by default" 🙄)
and triggers a quota rescan at the top of each call to `snapper list`.
(see: https://github.com/openSUSE/snapper/issues/629)

This adds the `--disable-used-space` to every call to `snapper list`,
opting out of the used-space display and consequently the quota rescan.